Decimal/float math problem on minimum trade exec.

+++ lorenzbot.py
@@ -1183,7 +1183,7 @@ if __name__ == '__main__':
                 # Check for sell conditions
                 if float(high_bid_actual) >= float(sell_price_target):
                     # Check if sell total greater than minimum allowed
-                    if float(sell_amount_current * sell_price_target) <= 0.0001:
+                    if (sell_amount_current * sell_price_target) <= Decimal(0.0001):
                         logger.warning('Trade total must be >= $0.0001. Skipping Trade.')
                         sell_skips += 1
                     else:
@@ -1193,7 +1193,7 @@ if __name__ == '__main__':
                 # Check for buy conditions
                 elif float(low_ask_actual) <= float(base_price):
                     # Check if buy total greater than minimum allowed
-                    if float(buy_amount_current * low_ask_actual) <= 0.0001:
+                    if (buy_amount_current * low_ask_actual) <= Decimal(0.0001):
                         logger.warning('Trade total must be >= $0.0001. Skipping Trade.')
                         buy_skips += 1
                     else:
